requirements.txt
fastapi==0.114.0
uvicorn==0.30.0
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
PyPDF2==3.0.1
python-docx==1.1.2
openai==1.45.0
starlette==0.38.2
